chore(agent-toolkit): move automations tools to automations-tools directory
Renames workflows-tools/ → automations-tools/ and list-workflows/ →
list-automations/ to match the tool naming convention (list_automations,
manage_automations, create_automation). Also removes the stale
publish_workflow description note that incorrectly referenced
manage_automations as a way to retrieve draft IDs.
