+++ packages/agent-toolkit/src/core/tools/platform-api-tools/automations-tools/index.ts
@@ -1,3 +1,3 @@
-export * from './list-workflows/list-workflows-tool';
+export * from './list-automations/list-automations-tool';
 export * from './manage-automations/manage-automations-tool';
 export * from './create-automation/create-automation-tool';
