+++ packages/agent-toolkit/src/core/tools/platform-api-tools/automations-tools/manage-automations/manage-automations-tool.ts
