+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workflow-builder-tools/publish-workflow/publish-workflow-tool.ts
@@ -52,7 +52,6 @@ Returns:
 - workflowObjectId: the workflow object ID (unchanged)
 - workflowLiveId: the new live version ID — this changes on every publish, so do not cache it
 
-Note: a new draft is created after each publish. To make further changes, retrieve the new draft ID first (e.g. via manage_automations), then call update_workflow.
 `;
   }
